refactor(babel-standalone): restructure and update for `react-native-reanimated@3.3.0`
